Handle paths with spaces in them appropriately
Turns out $BASH_COMMAND is the literal, unparsed command string,
including quotes and backslash escapes.  The only way to actually
find out what $0 is is to pass the string to another instance of
bash and extract the first argument.

This commit also switches from trapping 'err' to using $PROMPT_COMMAND
to execute the actual autocd, as trapping err seems to break sometimes.